+++ mobile/dequarantine/lib/logic/functions/auth/sign_out_functions.dart
