+++ mobile/dequarantine/lib/logic/functions/auth/sign_up_functions.dart
