+++ mobile/dequarantine/lib/logic/functions/general/get_events_functions.dart
